requirements.txt

aiogram==3.3.0
python-dotenv==1.0.0
asyncpg==0.29.0
SQLAlchemy==2.0.25
alembic==1.13.1
openpyxl==3.1.2
aiosmtplib==3.0.1
email-validator==2.1.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-asyncio==0.21.1
